+++ ServerManager.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#ifndef __SERVERMANAGER
+#define __SERVERMANAGER
+
+#include "server.h"
+
+class ServerManager {
+	Server ServerIns;
+
+
+
+
+};
+
+#endif // ~SERVERMANAGER
