--- UDP_Server/UDP_Server.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UDP_Server", "UDP_Server\UDP_Server.vcxproj", "{B8DA4B27-2A27-4116-952F-3D8BB409E80B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Debug|x64.ActiveCfg = Debug|x64
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Debug|x64.Build.0 = Debug|x64
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Debug|x86.ActiveCfg = Debug|Win32
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Debug|x86.Build.0 = Debug|Win32
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Release|x64.ActiveCfg = Release|x64
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Release|x64.Build.0 = Release|x64
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Release|x86.ActiveCfg = Release|Win32
-		{B8DA4B27-2A27-4116-952F-3D8BB409E80B}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
