--- UDP_Server/UDP_Server/main.cpp
@@ -1,8 +1,0 @@
-#include "server.h"
-
-int main(void) {
-	Server sevIns(8888);
-
-	sevIns.ConnectUDP();
-
-}
