+++ main.cpp
@@ -1,0 +1,8 @@
+#include "server.h"
+
+int main(void) {
+	Server sevIns(8888);
+
+	sevIns.ConnectUDP();
+
+}
